backend/requirements.txt
flask
flask-cors
numpy
scipy
sympy
scikit-image
matplotlib
Pillow
